conexion a redux
