+++ clients/src/redux/reducers/mainReducers.js
@@ -4,13 +4,14 @@ import instructorReducers from "./instructorReducers";
 import studentReducers from "./studentReducers";
 /*==== Auth Reducers ===== */
 import authReducers from "./authReducers";
+/* ==== Calendar reducers ==== */
+import calendarReducers from "./calendarReducers";
 /*==== Cohort-group-pair ===== */
 import cohortReducers from "./cohortReducers";
 import groupReducers from "./groupReducers";
 import pair_progReducers from "./pair-progReducers";
 /*==== Eamil ===== */
 import emailReducers from "./emailReducers";
-
 /* ==== Redux ==== */
 import { combineReducers } from "redux";
 
@@ -18,6 +19,7 @@ const mainReducers = combineReducers({
   admin: adminReducers,
   student: studentReducers,
   instructor: instructorReducers,
+  calendar: calendarReducers,
   auth: authReducers,
   email: emailReducers,
   cohort: cohortReducers,
